Add cached array clean-up
