+++ src/server/jobs/cache-cleanup.ts
@@ -3,6 +3,7 @@ import { redis, REDIS_KEYS } from '~/server/redis/client';
 import { CacheTTL } from '~/server/common/constants';
 import { mergeQueue } from '~/server/redis/queues';
 import { limitConcurrency } from '~/server/utils/concurrency-helpers';
+import * as caches from '~/server/redis/caches';
 
 export const cacheCleanup = createJob('cache-cleanup', '0 */1 * * *', async () => {
   // Clean rate limit keys
@@ -34,4 +35,7 @@ export const cacheCleanup = createJob('cache-cleanup', '0 */1 * * *', async () =
     await mergeQueue(key);
   });
   await limitConcurrency(mergeTasks, 3);
+
+  // Clean up old caches
+  for (const cache of Object.values(caches)) await cache.cleanup();
 });
